limit num of setpower
Limit the number of times we are setting power because basically we are trying to set the power hundreds of times per second.  This makes one call at the end of the state and then it is never made again. there are 11 edits that can easily be taken out and the old code is commented out.  We can check out the reddit for other solutions but this one looked most probable.