separate manual and levels
Both manual and levels are working, but while the arm is moving, the manual controls cannot be used.  Pay attention to alreadyMoving variable which will tell you when the arm is moving during the finite state machine.  God sorry i messed this up the first time, my brain has literally been going in circles for like the last 12 hours.